+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Create database (aka `db*`) objects for JSON files and a test `/api/bcat` route:
 
 * ...tbd a test route for /api/bcat/ to write sone data in books catalog
 * ...tbd init the `db` objects to JSON databases (.../booklabd/data/)
+* 250831-b `booklab.booklabd.__init__` imported working directories (created in prev change)
 * 250831-a `booklab.__init__` create constants: PACKAGE _ROOT, DATA_ROOT, CONF_ROOT
 * 250830-b make gunicorn config python module in `booklab.conf` module (as new one)
 * 250830-a config for for gunicorn in serving `booklabd` server web app. file is symlinked to project root to be correct executed with pdm run as tool script